♻ refactor: user service에서 level service 분리
